% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_networks.R
\name{add_pool}
\alias{add_pool}
\title{Add a pool to the network}
\usage{
add_pool(network, poolname, shape, unitnames, verbose = FALSE)
}
\arguments{
\item{network}{The network getting connected up, usually created by
read_network(), but probably new_network() in this case}

\item{poolname}{Name of the pool, needs to be a legal R variable name}

\item{shape}{A vector specifying the size of each dimension within the pool The numerical value of the connection weight}

\item{unitnames}{A vector of unitnames, one for each unit in the pool. These
need to be legal R variable names. Usually they would be generated automatically
using the gen_unitnames() function}

\item{verbose}{Gives details about the units as they are being added, useful}
}
\value{
A network with the new pool
}
\description{
Used internally by read_network().
\strong{Not needed} if you are using read_network() to create your network.
These and other functions are only used if you are creating your network
direct from code. Know that when a network
is read from .yaml file, it uses these routines so you don't have to.
}
