% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_cycle.R
\name{clear_external}
\alias{clear_external}
\title{Zero the external input to the network.}
\usage{
clear_external(network)
}
\arguments{
\item{network}{The network}
}
\value{
The network without external input
}
\description{
This is strictly to make intentions clear. It can be done as easily as:
network$external = 0
}
\seealso{
read_net(), cycle(), reset(), set_external()
}
