% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_cycle.R
\name{cycle}
\alias{cycle}
\title{Run the network}
\usage{
cycle(network, ncycles = 1, warnings = TRUE, verbose = FALSE, OG = FALSE)
}
\arguments{
\item{network}{A network, usually created with \code{read_net()}}

\item{ncycles}{The number of cycles to run (default = 1)}

\item{warnings}{If TRUE, will give a warning if there's no external input (default TRUE)}

\item{verbose}{Prints the cycle number and current sum-squared activation
difference from the previous cycle}

\item{OG}{If TRUE, will run network cycles using loop-based code based on
McClelland's original PDP Handbook, link below. Otherwise runs a vectorised version which
is not optimised but much faster, especially with networks of 100+ units
(default FALSE)}
}
\value{
A network with updated activation state and activation log,
suitable for \code{plot_log()}
}
\description{
The network is run in discrete cycles. Each cycle, the input for all units
is calculated, and then activation values for all units are updated.
}
\details{
The first step of a cycle
is determining the \emph{input} coming in to each unit. The input to a
unit is based on the weighted activity of units connecting to it, and
may optionally include random noise.

The second step of the cycle is to \emph{update} unit activations based on their
input. The activation function is direct from McClelland & Rumelhart (1981).
}
\seealso{
read_net(), set_external(), plot_log(), reset()
}
