% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_dsl7.R
\name{iac.cycle}
\alias{iac.cycle}
\title{Cycle -- run the network}
\usage{
iac.cycle(state.act, weights, external.act, paramlist)
}
\arguments{
\item{state.act}{A vector of current unit activations (length=net$nunits)}

\item{weights}{Weight matrix, normally net$wt.matrix}

\item{external.act}{A vector of external input (length=net$nunits)}

\item{paramlist}{A list(noise=x, alpha=x, gamma=x, ext.weight=x, rest=x, min=x, max=x, decay=x).Normally read this in from a file with iac.readparams()}
}
\description{
Run one cycle of the network
Usually you would do this after creating the network, eg net=iac.readfile('dualroutes.yaml')
}
\examples{
state.act = iac.cycle(state.act, net$wt.matrix, external.act, params)
log.act = iac.log.update(log.act, state.act)
}
\keyword{network}
\keyword{run}
