% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_dsl7.R
\name{iac.log.update}
\alias{iac.log.update}
\title{Update the log of activations}
\usage{
iac.log.update(current.log, new.state.act)
}
\arguments{
\item{log.act}{A matrix cols=net$nunits, rows=activations on each cycle}

\item{state.act}{The activations just computed from iac.cycle}
}
\description{
Usually you would do this after every cycle of the network
}
\examples{
state.act = iac.cycle(state.act, net$wt.matrix, external.act, .4, params)
log.act = iac.log.update(log.act, state.act)
}
\keyword{network}
\keyword{run}
