% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_dsl7.R
\name{iac.readnet}
\alias{iac.readnet}
\title{Read the network specificaton from a YAML file}
\usage{
iac.readnet(yaml_file, verbose = FALSE)
}
\arguments{
\item{yaml_file}{Name of yaml file}

\item{verbose}{If true, then pools and connections are printed to the screen (default=FALSE)}
}
\description{
Details about how to setup the yaml file forthcoming
}
\examples{
iac.readnet('dualroutes.yaml', verbose = TRUE)

}
\keyword{network}
\keyword{read}
