% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_dsl7.R
\name{iac.readparams}
\alias{iac.readparams}
\title{Read the network parameters from a YAML file
Expected to specify the following (example values shown):}
\usage{
iac.readparams(yaml_file)
}
\arguments{
\item{yaml_file}{Name of yaml file}
}
\description{
Read the network parameters from a YAML file
Expected to specify the following (example values shown):
}
\examples{
iac.readparams('params.yaml')

}
\keyword{network}
\keyword{read}
