% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_utils.R
\name{iac_example}
\alias{iac_example}
\title{Get path to IAC example}
\usage{
iac_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If NULL all the example files are listed.}
}
\description{
iac comes bundled with some example files in its inst/extdata directory. This
function make them easy to access.
}
\examples{
# returns a list of all available files
iac_example()
# load the jets-sharks network that comes with the iac package
jetsh = iac_example("jets_sharks.yaml")
net = read_net(jetsh)
# open an example network for inspection in RStudio
if (interactive()) file.edit(iac_example("what_where.yaml"))
}
