% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_utils.R
\name{plot_activation}
\alias{plot_activation}
\title{Plot selected unit activations by cycle}
\usage{
plot_activation(toplot, roi = NULL, cycles = NULL, mar = c(4, 4, 2, 3), ...)
}
\arguments{
\item{toplot}{Usually a network -- network$log is then used for plotting.
Alternatively, a dataframe of unit activations, with a column named 'cycle',
can be used. If "cycle" is not present, a vector from 1 to nrow of the
data.frame is used.}

\item{roi}{A vector specifying which columns to plot, this can be either
numeric values specifying indices, or ideally, unit names}

\item{cycles}{An optional vector of cycle numbers to be plotted. The default
is to plot all cycles,}

\item{mar}{Optionally call the mar graphics parameter, useful for making more
room for labels on the right (default = c(4,4,2,3))}

\item{...}{
  Arguments passed on to \code{\link[graphics:matplot]{graphics::matplot}}
  \describe{
    \item{\code{x,y}}{vectors or matrices of data for plotting.  The number of
    rows should match.  If one of them are missing, the other is taken
    as \code{y} and an \code{x} vector of \code{1:n} is used.
    Missing values (\code{NA}s) are allowed. Typically,
    \code{\link{class}(.)}es of \code{x} and \code{y} such as
    \code{"\link{Date}"} are preserved.}
    \item{\code{type}}{character string (length 1 vector) or vector of 1-character
    strings indicating the type of plot for each
    column of \code{y}, see \code{\link[graphics]{plot}} for all possible
    \code{type}s.  The first character of \code{type}
    defines the first plot, the second character the second, etc.
    Characters in \code{type} are cycled through; e.g., \code{"pl"}
    alternately plots points and lines.}
    \item{\code{lty,lwd,lend}}{vector of line types, widths, and end styles.
    The first element is for the first column, the second element for
    the second column, etc., even if lines are not plotted for all
    columns. Line types will be used cyclically until all plots are
    drawn.}
    \item{\code{pch}}{character string or vector of 1-characters or integers for
    plotting characters, see \code{\link[graphics]{points}} for details.
    The first character is the plotting-character for the first plot,
    the second for the second, etc.  The default is the digits (1
    through 9, 0) then the lowercase and uppercase letters.
  }
    \item{\code{col}}{vector of colors.  Colors are used cyclically.}
    \item{\code{cex}}{vector of character expansion sizes, used cyclically.
    This works as a multiple of \code{\link[graphics]{par}("cex")}.  \code{NULL} is
    equivalent to \code{1.0}. }
    \item{\code{bg}}{vector of background (fill) colors for the open plot
    symbols given by \code{pch = 21:25} as in \code{\link[graphics]{points}}.  The
    default \code{NA} corresponds to the one of the underlying function
    \code{\link[graphics]{plot.xy}}.}
    \item{\code{xlab,ylab}}{titles for x and y axes, as in \code{\link[graphics]{plot}}.}
    \item{\code{xlim,ylim}}{ranges of x and y axes, as in \code{\link[graphics]{plot}}.}
    \item{\code{add}}{logical.  If \code{TRUE}, plots are added to current one,
    using \code{\link[graphics]{points}} and \code{\link[graphics]{lines}}.}
    \item{\code{verbose}}{logical.  If \code{TRUE}, write one line of what is
    done.}
  }}
}
\description{
Plot selected unit activations by cycle
}
