% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_utils.R
\name{plot_log}
\alias{plot_log}
\title{Plot a network log with selected unit activations}
\usage{
plot_log(toplot, roi = NULL, cycles = NULL, mar = c(4, 4, 2, 3), ...)
}
\arguments{
\item{toplot}{A network or network$log to be plotted. Alternatively, a
dataframe of unit activations, with a column named 'cycle', can be used. If
"cycle" is not present, a vector from 1 to nrow of the data.frame is used.}

\item{roi}{A vector specifying which columns to plot, this can be either
numeric values specifying indices, or ideally, unit names}

\item{cycles}{An optional vector of cycle numbers to be plotted. The default
is to plot all cycles,}

\item{mar}{Optionally call the mar graphics parameter, useful for making more
room for labels on the right (default = c(4,4,2,3))}

\item{...}{The various optional arguments for plot, eg xlim, ylab, main, etc}
}
\description{
Plot a network log with selected unit activations
}
\examples{
# load the jets-sharks example
jetsh = read_net(iac_example("jets_sharks.yaml"), verbose =FALSE)
# do a partial retrieval of Ken's properties from his name
jetsh = set_external(jetsh, "Ken", 1.0)
jetsh = iac::cycle(jetsh, ncycles = 100)
# Ken is a burglar in the Sharks, what is retrieved from his name?
# plot Ken's properties and some others
plot_log(jetsh, roi=c("Ken", "_Ken", "jets", "sharks", "burglar", "bookie"),
main="Ken is a burgling Shark")

}
