% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_networks.R
\name{read_net}
\alias{read_net}
\title{Create a network from a file}
\usage{
read_net(yaml_file, verbose = TRUE, network = NULL)
}
\arguments{
\item{yaml_file}{The .yaml file with the network description}

\item{verbose}{If TRUE (default) then information describing the network
is printed to the console as the network is being assembled}

\item{network}{This will normally be NULL. However, instead of reading a
.yaml file to create a new network, you can also read a .yaml file which
will add pools or connections to an existing network. In this case, the
existing network being elaborated is passed as an argument here.}
}
\value{
A new network as described in the .yaml file
}
\description{
This is the intended method for network creation. Details are given in the
document iac_creating_networks.md.
}
\examples{
jetsh = read_net(iac_example("jets_sharks.yaml"))
jetsh = set_external(jetsh, "Ken", 1.0)
jetsh = iac::cycle(jetsh, ncycles = 100)
#
plot_log(jetsh, roi=c("Ken", "_Ken", "jets", "sharks", "burglar", "bookie"),
main="Ken is a burgling Shark")

}
