% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_cycle.R
\name{set_external}
\alias{set_external}
\title{Set the external input to the network.}
\usage{
set_external(network, units, activations)
}
\arguments{
\item{network}{The network}

\item{units}{A vector specifying which units to receive external input.
This can be either numeric values specifying indices, or probably better,
a vector of unit names}

\item{activations}{A numeric vector giving the value of external input
for each of the specified units}
}
\value{
The network with external inputs updated for subsequent calls
to cycle()
}
\description{
This external input will remain
until it is cleared by clear_external, or until set_external is called again.
}
\seealso{
read_net(), cycle(), reset(), clear_external()
}
