% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_utils.R
\name{show_weights}
\alias{show_weights}
\title{Inspect the NON-ZERO weights from or to a unit or units}
\usage{
show_weights(network, from = NULL, to = NULL)
}
\arguments{
\item{network}{The network}

\item{from, to}{A vector of unitnames (not pool names). Exactly one of these
is specified. Currently, if both are, \code{to} is ignored.}
}
\value{
The vector of non-zero weights from or to the specified units.
}
\description{
A way to quickly check network connections
}
\examples{
# load the jets_sharks network
jetsh = read_net(iac_example("jets_sharks.yaml"), verbose = FALSE)
# see the non-zero connections from the Ken instance (_Ken)
show_weights(jetsh, from = '_Ken')

}
