% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_utils.R
\name{weights_slice}
\alias{weights_slice}
\title{Extract the weights between two pools}
\usage{
weights_slice(network, from, to)
}
\arguments{
\item{network}{The network}

\item{from, to}{The names of existing pools within network}
}
\value{
A matrix from network$weights matrix, where rows are the units in
the \code{from} pool, and columns are units from the \code{to} pool.
This matrix is suitable for plot_weights() or other inspection
}
\description{
Extract the weights between two pools
}
\seealso{
plot_weights(), read_net()
}
