% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_cycle.R
\name{sim_batch}
\alias{sim_batch}
\title{Runs a batch of simulations. This is useful when you want to generate a
distrbution of results when using the noise parameter (without noise, every
run of a network will be the same)}
\usage{
sim_batch(network, nsims, ncycles)
}
\arguments{
\item{network}{The network}

\item{nsims}{The number of simulations in this batch}

\item{ncycles}{The number of cycles for each sim}
}
\value{
A dataframe where columns represent sim number, cycle, and every unit
in the network, and rows provide activations for each unit on each cycle of
each sim.
}
\description{
Runs a batch of simulations. This is useful when you want to generate a
distrbution of results when using the noise parameter (without noise, every
run of a network will be the same)
}
\seealso{
read_net(), cycle(),
}
