% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac_utils.R
\name{weights_slice}
\alias{weights_slice}
\title{Extract the weights between two pools}
\usage{
weights_slice(network, from, to)
}
\arguments{
\item{network}{The network}

\item{from, to}{The names of existing pools within network}
}
\value{
A matrix from within \code{network$weights}, where rows are the units in
the \code{from} pool, and columns are units from the \code{to} pool.
This matrix is suitable for \code{plot_weights()} or other inspection
}
\description{
Extract the weights between two pools
}
\examples{
# load the jets_sharks network
jetsh = read_net(iac_example("jets_sharks.yaml"), verbose = FALSE)
weights_slice(jetsh, from='gang', to='instance')
}
\seealso{
plot_weights()
}
